Refactor: remove component folders #258
