+++ src/lib/components/organisms/Hero.svelte
