+++ src/lib/components/organisms/PillarsComponent.svelte
